app/requirements.txt
Django==4.2.11
django-redis==5.4.0
djangorestframework==3.14.0